restructure file naming and layout, initial tidy
