+++ src/components/Header/utils/navLinksHelper.js
@@ -1,4 +1,4 @@
-import theme from '../../utils/theme'
+import theme from '../../../utils/theme'
 
 const specialitiesMap = {
   engineering: [
